requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
langchain==0.1.16
langchain-groq==0.1.2
langgraph==0.0.39
python-dotenv==1.0.1
pydantic==2.5.3